[Jest] Migrate `v9.test.js` (#653)
